requirements.txt
Django==4.2.11
django-crispy-forms==2.4
crispy-bootstrap4==2024.1
django-mapbox-location-field==2.1.0
django-ckeditor==6.4.0
Pillow==9.5.0
django-betterforms==2.0.0

